requirements.txt
python-binance==1.0.19
pandas==2.1.4
numpy==1.26.2
matplotlib==3.8.2
seaborn==0.13.0
plotly==5.18.0
ta-lib==0.4.28
ccxt==4.2.8
websockets==12.0
python-dotenv==1.0.0
sqlalchemy==2.0.23
apscheduler==3.10.4
streamlit==1.29.0
cryptography>=46.0.0
aiohttp>=3.9.0
